refactor: update authentication service and add mock project data
- Modified the handleGitHubCallback and logout functions in auth.service.ts to include return types for better type safety.
- Introduced a new mock project data file (project.mock.ts) to provide sample project data for testing and development.
- Updated import paths in project service files to reflect the new mock data structure, ensuring consistency across the application.
